requirements.txt
Flask==1.1.1
Flask-MySQLdb==0.2.0
mysqlclient==1.4.6
pylint==2.4.4
python-dotenv==0.10.3
requests==2.22.0
urllib3==1.25.7